requirements.txt
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.1
streamlit==1.27.2
folium==0.14.0
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.0
streamlit-folium==0.13.0
joblib==1.3.2
branca>=0.6.0